Refactor Python files to allow path injections and create DVC pipeline
