--- src/__init__.py
@@ -1,3 +1,0 @@
-
-if __name__ == "__main__":
-    print("Hi")
