--- src/features/__init__.py
@@ -1,1 +1,0 @@
-from features.build_features import *
