--- src/models/__init__.py
@@ -1,2 +1,0 @@
-from models.train_model import *
-from models.predict_model import *
